BUILD: quic+h3: 32-bit compilation errors fixes
In GH #1760 (which is marked as being a feature), there were compilation
errors on MacOS which could be reproduced in Linux when building 32-bit code
(-m32 gcc option). Most of them were due to variables types mixing in QUIC_MIN macro
or using size_t type in place of uint64_t type.

Must be backported to 2.6.

+++ include/haproxy/quic_enc.h
@@ -89,7 +89,7 @@ static inline size_t quic_int_getsize(uint64_t val)
 }
 
 /* Returns the maximum integer which may be encoded with <size> bytes */
-static inline size_t quic_max_int_by_size(int size)
+static inline uint64_t quic_max_int_by_size(int size)
 {
 	switch (size) {
 	case 1:
